feat: [P4PU-219] assistance form validation and unit test (#106)
