+++ src/components/AssistanceForm/index.ts
@@ -1,0 +1,1 @@
+export { AssistanceForm } from './AssistanceForm';
